Extract reusable method.
