+++ src/main/scala/org/broadinstitute/monster/common/StorageIO.scala
@@ -4,63 +4,13 @@ import com.spotify.scio.ScioContext
 import com.spotify.scio.coders.Coder
 import com.spotify.scio.io.ClosedTap
 import com.spotify.scio.values.SCollection
-import io.circe.{Encoder, Json, JsonNumber, Printer}
+import io.circe.{Encoder, Printer}
 import io.circe.syntax._
-import ujson.{AstTransformer, StringRenderer}
+import ujson.StringRenderer
 import upack.Msg
-import upickle.core.Visitor
-
-import scala.collection.mutable.ArrayBuffer
 
 /** Utilities for reading/writing messages from/to storage. */
 object StorageIO {
-
-  /**
-    * Utility class which can map circe's JSON representation to any type with
-    * a upickle Visitor implementation.
-    *
-    * Mostly a copy-paste of ujson-circe's `CirceJson` object, but with the behavior
-    * of numeric conversions tweaked to preserve the distinction between longs and doubles.
-    *
-    * @see https://github.com/lihaoyi/upickle/blob/master/ujson/circe/src/ujson/circe/CirceJson.scala
-    */
-  val circeVisitor: AstTransformer[Json] = new AstTransformer[Json] {
-
-    override def transform[T](j: Json, f: Visitor[_, T]): T = j.fold(
-      f.visitNull(-1),
-      if (_) f.visitTrue(-1) else f.visitFalse(-1),
-      n => n.toLong.map(f.visitInt64(_, -1)).getOrElse(f.visitFloat64(n.toDouble, -1)),
-      f.visitString(_, -1),
-      arr => transformArray(f, arr),
-      obj => transformObject(f, obj.toList)
-    )
-
-    def visitArray(length: Int, index: Int) =
-      new AstArrVisitor[Vector](x => Json.arr(x: _*))
-
-    def visitObject(length: Int, index: Int) =
-      new AstObjVisitor[ArrayBuffer[(String, Json)]](vs => Json.obj(vs: _*))
-
-    def visitNull(index: Int): Json = Json.Null
-
-    def visitFalse(index: Int): Json = Json.False
-
-    def visitTrue(index: Int): Json = Json.True
-
-    def visitFloat64StringParts(
-      s: CharSequence,
-      decIndex: Int,
-      expIndex: Int,
-      index: Int
-    ): Json = Json.fromJsonNumber(
-      if (decIndex == -1 && expIndex == -1)
-        JsonNumber.fromIntegralStringUnsafe(s.toString)
-      else JsonNumber.fromDecimalStringUnsafe(s.toString)
-    )
-
-    def visitString(s: CharSequence, index: Int): Json = Json.fromString(s.toString)
-  }
-
   /**
     * JSON formatter used when writing outputs.
     *
@@ -86,15 +36,7 @@ object StorageIO {
   )(implicit coder: Coder[Msg]): SCollection[Msg] =
     context
       .textFile(filePattern)
-      .transform(s"Parse Messages: $description")(
-        _.map { line =>
-          val maybeParsed = io.circe.parser.parse(line)
-          maybeParsed.fold(
-            err => throw new Exception(s"Failed to parse input line as JSON: $line", err),
-            js => circeVisitor.transform(js, Msg)
-          )
-        }
-      )
+      .transform(s"Parse Messages: $description")(_.map(msg.parseJsonString))
 
   /**
     * Write unmodeled messages to storage for use by downstream components.
